Rescan user payments integration
