+++ src/actions/types.js
@@ -154,3 +154,6 @@ export const TOGGLE_CREDITS_PAYMENT_POLLING = "TOGGLE_CREDITS_PAYMENT_POLLING";
 export const TOGGLE_PROPOSAL_PAYMENT_RECEIVED = "TOGGLE_PROPOSAL_PAYMENT_RECEIVED";
 
 export const RESET_PAYWALL_INFO = "RESET_PAYWALL_INFO";
+
+export const REQUEST_RESCAN_USER_PAYMENTS = "REQUEST_RESCAN_USER_PAYMENTS";
+export const RECEIVE_RESCAN_USER_PAYMENTS = "RECEIVE_RESCAN_USER_PAYMENTS";
