+++ src/connectors/user.js
@@ -27,6 +27,8 @@ const userConnector = connect(
     loggedInAsUserId: sel.userid,
     user: sel.user,
     error: sel.apiUserError,
+    errorRescan: sel.apiRescanUserPaymentsError,
+    isLoadingRescan: sel.isApiRequestingRescanUserPayments,
     isLoading: or(sel.isApiRequestingUser, sel.isApiRequestingMe),
     isLoadingProposals: or(sel.userProposalsIsRequesting, sel.isApiRequestingPropsVoteStatus),
     isTestnet: sel.isTestNet,
@@ -77,7 +79,8 @@ const userConnector = connect(
     onIdentityImported: act.onIdentityImported,
     confirmWithModal,
     onUpdateUserKey,
-    onEditUser: act.onEditUser
+    onEditUser: act.onEditUser,
+    onRescan: act.onRescanUserPayments
   }, dispatch)
 );
