+++ src/reducers/api/index.js
@@ -1,0 +1,5 @@
+import api from "./api";
+export * from "./handlers";
+export * from "./api";
+
+export default api;
